Update UI to Web Function
