+++ package.json
@@ -17,7 +17,7 @@
     "react-pageflip": "^2.0.3",
     "react-pdf": "^5.7.1",
     "react-router-dom": "^6.26.2",
-    "react-scripts": "^5.0.0",
+    "react-scripts": "5.0.0",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
