+++ src/App.js
@@ -1,16 +1,16 @@
-import "./App.css";
-import './index.css';
+import "./styles/App.css";
+import './styles/index.css';
 import React from "react";
 import { BrowserRouter, Routes, Route, Navigate } from "react-router-dom";
 import Login from "./components/auth/Login.jsx";
 import Register from "./components/auth/Register.jsx";
-import Header from "./components/Header.jsx";
-import Home from "./components/Homepage.jsx";
-import FlipBook from "./components/Flipbook.jsx";
-import Share from "./components/Share"; // Import the Share component
+import Header from "./components/layout/Header.jsx";
+import Home from "./pages/Homepage/Homepage.jsx";
+import FlipBook from "./pages/Flipbook/Flipbook.jsx";
+import Share from "./pages/Share/Share.jsx"; // Import the Share component
 import { AuthProvider } from "./contexts/authContext.jsx";
 import { PdfProvider } from "./contexts/PdfContext.jsx";
-import ProtectedRoute from "./components/ProtectedRoute.jsx";
+import ProtectedRoute from "./components/common/ProtectedRoute.jsx";
 
 function App() {
   return (
