+++ src/components/layout/Header.jsx
@@ -1,9 +1,8 @@
 import React from "react";
 import { Link, useNavigate } from "react-router-dom";
-import { useAuth } from "../contexts/authContext";
-import { doSignOut } from "../firebase/auth";
+import { useAuth } from "../../contexts/authContext";
+import { doSignOut } from "../../firebase/auth";
 import { FaBook } from "react-icons/fa";
-
 
 const Header = () => {
   const navigate = useNavigate();
