+++ src/styles/Modal.css
@@ -96,7 +96,6 @@
   align-items: center;
   z-index: 1000;
 }
-
 .modal-content {
   background: #fff;
   /* Nền trắng cho modal */
